:truck: Move wedding code to its own folder
